Safari-611.1.10/LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-es-MX.html

<!DOCTYPE html><!-- webkit-test-runner [ language=es-MX ] -->
<html>
<head>
<script src="../../../../../resources/js-test.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "es-mx");
</script>
</body>
</html>
