Safari-611.1.10/LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-es-MX-expected.txt

FAIL navigator.language should be es-mx. Was es-xl.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
